+++ CHANGELOG.md
@@ -6,8 +6,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Compound Coordinate Transforms through an intermediary CRS
+
 ### Fixed
 - Adjustment to OSGB36 datum transform e.g. EPSG: 27700
+- WGS84 intermediary transformations in non geographic CRS pairs e.g. EPSG: 27700 [#32](https://github.com/locationtech/proj4j/issues/32)
 - +nadgrids=@null support e.g. EPSG: 3857
 
 ## [1.1.3] - 2021-06-17